The serialized origin is asked of the module that owns URIs
`is_valid_serialized_origin` was in `headers.rs` for being reachable from a
header field. Two of its neighbours — `extract_origin_if_absolute` and
`validate_origin_value` — were already in `uri.rs`, and `uri.rs` was reaching
back across for it. A helper shelved by the field that carries it rather than
by what it asks makes exactly that shape.

Four tests came with it, including the rstest table, whose `#[case]` rows sit
above the function and are easy to strand: the doc paragraphs explaining the
table had already drifted apart from it in the old file, and are reattached
here. Verified by name — `invalid_serialized_origin_cases::case_1..3` and
`test_is_valid_serialized_origin` now run under `helpers::uri`, and the crate
total is unchanged at 6032.

headers.rs: 2344 → 2064 non-test lines across the three moves so far.
